requirements-dev.txt
asgiref==3.4.1
attrs==22.2.0
autopep8==2.0.2
black==23.1.0
bleach==6.0.0
boto3==1.28.30
botocore==1.31.30
certifi==2022.12.7
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.8
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.12
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.13.0
django-advanced-filters==2.0.0
django-app-novadata==0.1.1
django-appconf==1.0.5
django-avatar==7.1.1
django-browser-reload==1.7.0
django-colorfield==0.8.0
django-crum==0.7.9
django-filter==22.1
django-import-export==3.1.0
django-object-actions==4.1.0
django-quill-editor==0.1.40
django-storages==1.13.2
django-vite==2.1.3
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
dnspython==2.3.0
docutils==0.19
drf-spectacular==0.26.0
drf-spectacular-sidecar==2023.3.1
et-xmlfile==1.1.0
filelock==3.15.4
flake8==6.0.0
identify==2.6.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.0.0
inflection==0.5.1
jaraco.classes==3.2.3
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.17.3
keyring==23.13.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupPy==1.14
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
mypy==1.0.1
mypy-extensions==1.0.0
nodeenv==1.9.1
novadata_utils==0.3.7
odfpy==1.4.1
openpyxl==3.1.1
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pre-commit==3.7.1
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pywin32-ctypes==0.2.2
PyYAML==6.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.2
s3transfer==0.6.2
SecretStorage==3.3.3
simplejson==3.18.3
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
static3==0.7.0
tablib==3.3.0
tomli==2.0.1
twine==4.0.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.26.3
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
